Use SIZE_T for SectionSize in RegisterFrozenSegment (dotnet/coreclr#25044)
Commit migrated from https://github.com/dotnet/coreclr/commit/dd9e417dc0f11d7ec3bed54de75853ecf87e70e5

+++ src/coreclr/src/vm/comutilnative.h
@@ -146,7 +146,7 @@ public:
 
 #ifdef FEATURE_BASICFREEZE
     static
-    void* QCALLTYPE RegisterFrozenSegment(void *pSection, INT32 sizeSection);
+    void* QCALLTYPE RegisterFrozenSegment(void *pSection, SIZE_T sizeSection);
 
     static
     void QCALLTYPE UnregisterFrozenSegment(void *segmentHandle);
